+++ Krowi_AchievementFilter.lua
@@ -66,6 +66,8 @@ function loadHelper:OnEvent(event, arg1, arg2)
 
             addon.Icon.Load();
             addon.Tutorials.Load();
+
+            addon.UnitData.Load(); -- Might be moved to PLAYER_LOGIN event but easier for testing on every /reload
         elseif arg1 == "Blizzard_AchievementUI" then -- This needs the Blizzard_AchievementUI addon available to load
             addon.Data.Load();