+++ Localization/frFR.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-10-24 19-19-51 ]] --
+-- [[ Exported at 2022-11-01 14-17-28 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -17,4 +17,22 @@ addon.ExtraWrathStrings.frFR.Load(L);
 addon.GlobalStrings.Load(L);
 addon.Plugins:LoadLocalization(L);
 
+L["Achievement points earned by"] = "Points de Hauts faits obtenus par";
+L["Achievements"] = "Hauts faits";
+L["Expansions"] = "Extensions";
+L["Events"] = "Évènements";
+L["PvP"] = "Joueur contre Joueur";
+L["Specials"] = "Spécifique";
+L["Merge Small Categories"] = "Fusionner les catégories mineures";
+L["Obtainable"] = "Obtenable";
+L["Not Obtainable"] = "Non Obtenable";
+L["Covenant"] = "Congrégation";
+L["ID"] = "ID";
+L["Earned By"] = "Obtenu par";
+L["Account"] = "Compte";
+L["Character only"] = "Personnage uniquement";
+L["Tabs"] = "Onglets";
+L["Special Categories"] = "Catégorie Spéciale";
+L["Current Zone"] = "Zone Actuelle";
+L["Selected Zone"] = "Zone Sélectionnée";